Fix script path and update to run with bash
Updated the path in `run-integration.sh` to use the correct relative
path for `wait-for-it.sh`.

By default, GitHub Actions uses the `sh` shell to execute scripts,
which leads to syntax errors when running the `wait-for-it.sh` script
because it contains Bash-specific syntax.
Changed the execution of `wait-for-it.sh` to use `bash` instead of `sh`
to avoid these syntax errors.